requirements.txt
kivy
kivymd
python-dotenv
sqlalchemy
mysql-connector-python
alembic
validators
passlib[bcrypt]